add config path persistent

+++ micro-cluster/datatransfer/importexport/manager.go
@@ -113,6 +113,7 @@ func (mgr *ImportExportManager) ExportData(ctx context.Context, request message.
 		ClusterID:       request.ClusterID,
 		TransportType:   string(constants.TransportTypeExport),
 		FilePath:        mgr.getDataExportFilePath(&request, exportDir, true),
+		ConfigPath:      exportDir,
 		ZipName:         request.ZipName,
 		StorageType:     request.StorageType,
 		Comment:         request.Comment,
@@ -217,6 +218,7 @@ func (mgr *ImportExportManager) ImportData(ctx context.Context, request message.
 			ClusterID:       request.ClusterID,
 			TransportType:   string(constants.TransportTypeImport),
 			FilePath:        mgr.getDataImportFilePath(&request, importDir, true),
+			ConfigPath:      importDir,
 			ZipName:         constants.DefaultZipName,
 			StorageType:     request.StorageType,
 			Comment:         request.Comment,
@@ -259,6 +261,7 @@ func (mgr *ImportExportManager) ImportData(ctx context.Context, request message.
 			ClusterID:       request.ClusterID,
 			TransportType:   string(constants.TransportTypeImport),
 			FilePath:        mgr.getDataImportFilePath(&request, importDir, true),
+			ConfigPath:      importDir,
 			ZipName:         constants.DefaultZipName,
 			StorageType:     request.StorageType,
 			Comment:         request.Comment,
@@ -362,8 +365,21 @@ func (mgr *ImportExportManager) DeleteDataTransportRecord(ctx context.Context, r
 	if string(constants.StorageTypeS3) != record.StorageType {
 		filePath := filepath.Dir(record.FilePath)
 		go func() {
-			removeErr := os.RemoveAll(filePath)
-			framework.LogWithContext(ctx).Infof("remove file path %s, record: %+v, error: %v", filePath, record, removeErr)
+			if filePath != "" {
+				removeErr := os.RemoveAll(filePath)
+				framework.LogWithContext(ctx).Infof("remove file path %s, record: %+v, error: %v", filePath, record, removeErr)
+			}
+			if record.ConfigPath != "" {
+				removeErr := os.RemoveAll(record.ConfigPath)
+				framework.LogWithContext(ctx).Infof("remove config path %s, record: %+v, error: %v", record.ConfigPath, record, removeErr)
+			}
+		}()
+	} else {
+		go func() {
+			if record.ConfigPath != "" {
+				removeErr := os.RemoveAll(record.ConfigPath)
+				framework.LogWithContext(ctx).Infof("remove config path %s, record: %+v, error: %v", record.ConfigPath, record, removeErr)
+			}
 		}()
 	}
